refactor: extract display/preset handlers to handlers.rs (#165)
