+++ iem-mixer/iem-ui/src/pages/mixer/mod.rs
@@ -12,13 +12,14 @@ use crate::components::category_tabs::{Category, CategoryTabs};
 use crate::components::eq_modal::EQModal;
 use crate::components::limiter_modal::LimiterModal;
 use crate::components::pin_change_modal::PinChangeModal;
-use crate::components::preset_modal::{ChannelState, PresetData, PresetModal};
+use crate::components::preset_modal::PresetModal;
 use crate::components::settings_modal::SettingsModal;
 use crate::components::snapshot_modal::SnapshotModal;
 use crate::components::toolbar::Toolbar;
 
 mod components;
 mod connection;
+mod handlers;
 mod helpers;
 mod push;
 mod state;
@@ -147,205 +148,19 @@ pub fn MixerPage() -> impl IntoView {
 
     // Process channels for display (handle stereo pairs, pin/hide)
     // Memoized to avoid recomputation on every meter update
-    let display_channels = Memo::new(move |_| {
-        let chs = channels.get();
-        let member = member_id();
-        let my_input = format!("{} MIC", member.to_uppercase());
-        let active_cat = active_category.get();
-        let pinned = pinned_channels.get();
-        let hidden = hidden_channels.get();
-
-        let mut result = Vec::new();
-        let mut seen_pairs: std::collections::HashSet<String> = std::collections::HashSet::new();
-
-        for ch in &chs {
-            if ch.stereo_side.as_deref() == Some("R") {
-                continue;
-            }
-
-            let is_my_input = ch.name.to_uppercase() == my_input;
-            let is_pinned = pinned.contains(&ch.track_index);
-            let is_hidden = hidden.contains(&ch.track_index);
-
-            // Hidden tab: only show hidden channels
-            if active_cat == Category::Hidden {
-                if !is_hidden {
-                    continue;
-                }
-            } else if active_cat == Category::Main {
-                // Main tab: show "me" channel + pinned channels
-                // Hidden does NOT remove pinned channels from Main — only from category tabs
-                if !is_my_input && !is_pinned {
-                    continue;
-                }
-            } else {
-                // Category tabs: filter by category, skip hidden
-                if !active_cat.matches(&ch.category) {
-                    continue;
-                }
-                if is_hidden {
-                    continue;
-                }
-            }
-
-            let (is_stereo, partner_index) = if ch.stereo_side.as_deref() == Some("L") {
-                if let Some(ref pair_name) = ch.stereo_pair {
-                    if seen_pairs.contains(pair_name) {
-                        continue;
-                    }
-                    seen_pairs.insert(pair_name.clone());
-                    let partner = chs
-                        .iter()
-                        .find(|c| {
-                            c.stereo_pair.as_ref() == Some(pair_name)
-                                && c.stereo_side.as_deref() == Some("R")
-                        })
-                        .map(|c| c.track_index);
-                    (true, partner)
-                } else {
-                    (false, None)
-                }
-            } else {
-                (false, None)
-            };
-
-            let display_name = if is_stereo {
-                ch.name.trim_end_matches(" L").to_string()
-            } else {
-                ch.name.clone()
-            };
-
-            result.push(DisplayChannel {
-                track_index: ch.track_index,
-                display_name,
-                is_stereo,
-                partner_index,
-                is_my_input,
-            });
-        }
-
-        // Sort Stems: Click first, Guide second, then rest
-        if active_cat == Category::Stems {
-            result.sort_by_key(|ch| {
-                let name_upper = ch.display_name.to_uppercase();
-                if name_upper == "CLICK" {
-                    0
-                } else if name_upper == "GUIDE" {
-                    1
-                } else {
-                    2
-                }
-            });
-        }
-
-        // Sort Main: own channel first, pinned channels after
-        if active_cat == Category::Main {
-            result.sort_by_key(|ch| if ch.is_my_input { 0 } else { 1 });
-        }
-
-        result
-    });
+    let display_channels = handlers::make_display_channels(
+        channels,
+        member_id.clone(),
+        active_category,
+        pinned_channels,
+        hidden_channels,
+    );
 
     // Preset handlers
-    let get_current_state = Callback::new(move |_: ()| {
-        let chs = channels.get();
-        let mut channel_states = HashMap::new();
-        for ch in &chs {
-            channel_states.insert(
-                ch.track_index,
-                ChannelState {
-                    vol: ch.level_db,
-                    mute: ch.muted,
-                    pan: ch.pan,
-                },
-            );
-        }
-        let current_stems_level = if stems_bus_idx.get_untracked().is_some() {
-            Some(stems_level.get_untracked())
-        } else {
-            None
-        };
-        // Include cached EQ data if the user has loaded EQ for a track
-        let eq_data = {
-            let bands = eq_bands.get_untracked();
-            let open = eq_open.get_untracked();
-            if !bands.is_empty() {
-                if let Some((track_idx, _)) = open {
-                    let mut eq_map = HashMap::new();
-                    eq_map.insert(
-                        track_idx,
-                        bands
-                            .iter()
-                            .map(|b| crate::components::preset_modal::EqBandPreset {
-                                band_type: b.band_type.clone(),
-                                freq_hz: b.freq_hz,
-                                gain_db: b.gain_db,
-                                bw: b.bw,
-                                freq_norm: b.freq_norm,
-                                gain_norm: b.gain_norm,
-                                bw_norm: b.bw_norm,
-                            })
-                            .collect(),
-                    );
-                    Some(eq_map)
-                } else {
-                    None
-                }
-            } else {
-                None
-            }
-        };
-        PresetData {
-            channels: channel_states,
-            created_at: None,
-            updated_at: None,
-            stems_level_db: current_stems_level,
-            eq_bands: eq_data,
-        }
-    });
+    let get_current_state =
+        handlers::make_get_current_state(channels, stems_bus_idx, stems_level, eq_bands, eq_open);
 
-    let on_load_preset = Callback::new(move |preset: PresetData| {
-        if !connected.get() {
-            web_sys::console::warn_1(&"Preset loading blocked: not connected to REAPER".into());
-            return;
-        }
-
-        // Update local state
-        let _ = set_channels.try_update(|chs| {
-            for ch in chs.iter_mut() {
-                if let Some(state) = preset.channels.get(&ch.track_index) {
-                    ch.level_db = state.vol;
-                    ch.muted = state.mute;
-                    ch.pan = state.pan;
-                }
-            }
-        });
-
-        // Send all changes via WebSocket
-        for (track_index, state) in &preset.channels {
-            ws_send(
-                ws,
-                &iem_core::ClientMsg::SetLevel {
-                    track_index: *track_index,
-                    level_db: state.vol,
-                },
-            );
-            ws_send(
-                ws,
-                &iem_core::ClientMsg::SetMute {
-                    track_index: *track_index,
-                    muted: state.mute,
-                },
-            );
-            ws_send(
-                ws,
-                &iem_core::ClientMsg::SetPan {
-                    track_index: *track_index,
-                    pan: state.pan,
-                },
-            );
-        }
-    });
+    let on_load_preset = handlers::make_on_load_preset(connected, set_channels, ws);
 
     // Toolbar callbacks
     // Show engineer toolbar (Mute All + Listen) on any page when logged in as engineer
@@ -366,15 +181,7 @@ pub fn MixerPage() -> impl IntoView {
         let _ = set_snapshot_modal_visible.try_set(true);
     });
 
-    let mute_all_member = member_id.clone();
-    let on_mute_all = Callback::new(move |_: ()| {
-        let member = mute_all_member();
-        wasm_bindgen_futures::spawn_local(async move {
-            if let Err(e) = crate::api::batch_mute_all(&member).await {
-                web_sys::console::error_1(&format!("Mute all failed: {}", e).into());
-            }
-        });
-    });
+    let on_mute_all = handlers::make_on_mute_all(member_id.clone());
 
     let on_close_modal = Callback::new(move |_: ()| {
         let _ = set_preset_modal_visible.try_set(false);
